+++ tests/smoke-github-merge-pull-request.php
@@ -113,9 +113,6 @@ namespace DataMachineCode\GitHub {
     class PrReviewEscalationPolicy
     {
     }
-    class PrHomeboyReviewRunner
-    {
-    }
 }
 
 namespace {